+++ rbx_reflection_database/src/lib.rs
@@ -4,7 +4,7 @@ static ENCODED_DATABASE: &[u8] = include_bytes!("../database.msgpack");
 
 lazy_static::lazy_static! {
     static ref DATABASE: ReflectionDatabase<'static> = {
-        rmp_serde::decode::from_slice(ENCODED_DATABASE).unwrap()
+        rmp_serde::decode::from_slice(ENCODED_DATABASE).unwrap_or_else(|e| panic!("could not decode reflection database because: {}", e))
     };
 }