+++ rbx_xml/src/deserializer.rs
@@ -8,7 +8,7 @@ use rbx_dom_weak::{
     types::{Ref, SharedString, Variant, VariantType},
     InstanceBuilder, WeakDom,
 };
-use rbx_reflection::DataType;
+use rbx_reflection::{DataType, PropertyKind, PropertySerialization};
 
 use crate::{
     conversion::ConvertVariant,
@@ -607,7 +607,34 @@ fn deserialize_properties<R: Read>(
                 }
             };
 
-            props.insert(descriptor.name.to_string(), value);
+            match &descriptor.kind {
+                PropertyKind::Canonical {
+                    serialization: PropertySerialization::Migrate(migration),
+                } => {
+                    let new_property_name = &migration.new_property_name;
+                    let old_property_name = &descriptor.name;
+
+                    if !props.contains_key(new_property_name) {
+                        log::trace!(
+                            "Attempting to migrate property {old_property_name} to {new_property_name}"
+                        );
+                        match migration.perform(&value) {
+                            Ok(migrated_value) => {
+                                props.insert(new_property_name.to_string(), migrated_value);
+                                log::trace!(
+                                    "Successfully migrated property {old_property_name} to {new_property_name}"
+                                );
+                            }
+                            Err(error) => {
+                                return Err(reader.error(DecodeErrorKind::MigrationError(error)));
+                            }
+                        }
+                    }
+                }
+                _ => {
+                    props.insert(descriptor.name.to_string(), value);
+                }
+            };
         } else {
             match state.options.property_behavior {
                 DecodePropertyBehavior::IgnoreUnknown => {