+++ rbx_binary/src/tests/models.rs
@@ -63,4 +63,5 @@ binary_tests! {
     text_label_with_font,
     gui_inset_and_font_migration,
     folder_with_cframe_attributes,
+    folder_with_font_attribute,
 }