+++ rbx_xml/src/tests/models.rs
@@ -64,4 +64,5 @@ model_tests! {
     text_label_with_font,
     gui_inset_and_font_migration,
     folder_with_cframe_attributes,
+    folder_with_font_attribute,
 }