+++ rbx_types/CHANGELOG.md
@@ -5,11 +5,13 @@
 * Changed `BinaryString`'s non-human readable serde implementation to be identical to `Vec<u8>`. ([#276])
 * Added `Font::new` and `Font::regular` constructors. ([#283])
 * Added support for `CFrame` values in attributes. ([#296])
+* Added support for `Font` values in attributes. ([#299])
 
 [#271]: https://github.com/rojo-rbx/rbx-dom/pull/271
 [#276]: https://github.com/rojo-rbx/rbx-dom/pull/276
 [#283]: https://github.com/rojo-rbx/rbx-dom/pull/283
 [#296]: https://github.com/rojo-rbx/rbx-dom/pull/296
+[#299]: https://github.com/rojo-rbx/rbx-dom/pull/299
 
 ## 1.5.0 (2023-04-22)
 * Implemented `Font`. ([#248])